Refactor C code of test contracts
