+++ test/contracts/elrond/types.h
@@ -1,3 +1,8 @@
+#ifndef _TYPES_H_
+#define _TYPES_H_
+
 typedef unsigned char byte;
 typedef unsigned int u32;
 typedef unsigned long long u64;
+
+#endif
