+++ test/contracts/exec-dest-ctx-child/exec-dest-ctx-child.c
@@ -1,5 +1,6 @@
 #include "../elrond/context.h"
 #include "../elrond/bigInt.h"
+#include "../elrond/test_utils.h"
 
 byte dataA[20] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
 byte dataB[20] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
@@ -11,24 +12,8 @@ byte childKey[] =  "childKey........................";
 byte childData[] = "childData";
 byte childFinish[] = "childFinish";
 
-byte recipient[32]     = "childTransferReceiver...........";
+byte recipient[32] = "childTransferReceiver...........";
 byte value[] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,12};
-
-void not_ok() {
-	byte msg[] = "not ok";
-	finish(msg, 6);
-}
-
-void didCallerPay() {
-	bigInt bigInt_payment = bigIntNew(0);
-	bigIntGetCallValue(bigInt_payment);
-
-	long long payment = bigIntGetInt64(bigInt_payment);
-	if (payment != 99) {
-		byte message[] = "child execution requires tx value of 99";
-		signalError(message, 39);
-	}
-}
 
 void childFunction() {
 	int numArgs = getNumArguments();
@@ -37,7 +22,7 @@ void childFunction() {
 		signalError(message, 25);
 	}
 
-	didCallerPay();
+	didCallerPay(99);
 
 	// This transfer will appear alongside the transfers made by the parent.
 	byte transferData[100];
@@ -62,7 +47,7 @@ void childFunction_BigInts() {
 		signalError(message, 25);
 	}
 
-	didCallerPay();
+	didCallerPay(99);
 
 	int status = 0;
 
@@ -121,7 +106,7 @@ void childFunction_OutOfGas() {
 		signalError(message, 25);
 	}
 
-	didCallerPay();
+	didCallerPay(99);
 
 	storageStore(childKey, childData, 9);
 	finish(childFinish, 11);
