+++ test/contracts/exec-same-ctx-recursive/exec-same-ctx-recursive.c
@@ -1,24 +1,18 @@
 #include "../elrond/context.h"
 #include "../elrond/bigInt.h"
 #include "../elrond/types.h"
+#include "../elrond/test_utils.h"
+
+u64 maxGasForCalls = 100000;
 
 byte selfAddress[] = "parentSC........................";
 byte executeValue[] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,5};
 byte arguments[1] = {0};
 int argumentsLengths[1] = {1};
 
-byte recursiveIterationCounterKey[] = "recursiveIterationCounter.......";
-byte recursiveIterationBigCounterKey[] = "recursiveIterationBigCounter....";
-bigInt bigIntIterationCounterID = 16;
-
-void intTo3String(int value, byte *string, int startPos);
-void finishResult(int result);
-void incrementIterCounter();
-void incrementBigIntCounter();
-void storeIterationNumber(byte iteration, byte prefix);
-void finishIterationNumber(byte iteration, byte prefix);
-
-void recursiveMethodA();
+byte smallCounterKey[] = "recursiveIterationCounter.......";
+byte bigIntCounterKey[] = "recursiveIterationBigCounter....";
+bigInt bigIntCounterID = 16;
 
 void callRecursive() {
 	int numArgs = getNumArguments();
@@ -31,15 +25,15 @@ void callRecursive() {
 
   finishIterationNumber(iteration, 'R');
   storeIterationNumber(iteration, 'R');
-  incrementIterCounter();
-  incrementBigIntCounter();
+  incrementIterCounter(smallCounterKey);
+  incrementBigIntCounter(bigIntCounterID);
 
   // Run next iteration.
   byte functionName[] = "callRecursive";
   if (iteration > 0) {
     arguments[0] = iteration - 1;
     int result = executeOnSameContext(
-        10000,
+        maxGasForCalls,
         selfAddress,
         executeValue,
         functionName,
@@ -51,7 +45,7 @@ void callRecursive() {
 
     finishResult(result);
   } else {
-    bigIntStorageStoreUnsigned(recursiveIterationBigCounterKey, bigIntIterationCounterID);
+    bigIntStorageStoreUnsigned(bigIntCounterKey, bigIntCounterID);
   }
 }
 
@@ -75,7 +69,7 @@ void callRecursiveMutualMethods() {
   if (iteration > 0) {
     arguments[0] = iteration;
     int result = executeOnSameContext(
-        10000,
+        maxGasForCalls,
         selfAddress,
         executeValue,
         functionNameB,
@@ -96,14 +90,14 @@ void recursiveMethodA() {
 
   finishIterationNumber(iteration, 'A');
   storeIterationNumber(iteration, 'A');
-  incrementIterCounter();
-  incrementBigIntCounter();
+  incrementIterCounter(smallCounterKey);
+  incrementBigIntCounter(bigIntCounterID);
 
   byte functionNameB[] = "recursiveMethodB";
   if (iteration > 0) {
     arguments[0] = iteration - 1;
     int result = executeOnSameContext(
-        10000,
+        maxGasForCalls,
         selfAddress,
         executeValue,
         functionNameB,
@@ -114,9 +108,8 @@ void recursiveMethodA() {
     );
     finishResult(result);
   } else {
-    bigIntStorageStoreUnsigned(recursiveIterationBigCounterKey, bigIntIterationCounterID);
+    bigIntStorageStoreUnsigned(bigIntCounterKey, bigIntCounterID);
   }
-
 }
 
 void recursiveMethodB() {
@@ -124,14 +117,14 @@ void recursiveMethodB() {
 
   finishIterationNumber(iteration, 'B');
   storeIterationNumber(iteration, 'B');
-  incrementIterCounter();
-  incrementBigIntCounter();
+  incrementIterCounter(smallCounterKey);
+  incrementBigIntCounter(bigIntCounterID);
 
   byte functionNameB[] = "recursiveMethodA";
   if (iteration > 0) {
     arguments[0] = iteration - 1;
     int result = executeOnSameContext(
-        10000,
+        maxGasForCalls,
         selfAddress,
         executeValue,
         functionNameB,
@@ -142,63 +135,6 @@ void recursiveMethodB() {
     );
     finishResult(result);
   } else {
-    bigIntStorageStoreUnsigned(recursiveIterationBigCounterKey, bigIntIterationCounterID);
+    bigIntStorageStoreUnsigned(bigIntCounterKey, bigIntCounterID);
   }
-
-}
-
-void storeIterationNumber(byte iteration, byte prefix) {
-  byte keyIter[] = "XkeyNNN.........................";
-  byte valueIter[] = "XvalueNNN";
-  intTo3String(iteration, keyIter, 4);
-  intTo3String(iteration, valueIter, 6);
-  keyIter[0] = prefix;
-  valueIter[0] = prefix;
-  storageStore(keyIter, valueIter, 9);
-}
-
-void finishIterationNumber(byte iteration, byte prefix) {
-  byte finishIter[] = "XfinishNNN";
-  intTo3String(iteration, finishIter, 7);
-  finishIter[0] = prefix;
-  finish(finishIter, 10);
-}
-
-void intTo3String(int value, byte *string, int startPos) {
-  string[startPos + 2] = (byte)('0' + value % 10);
-  string[startPos + 1] = (byte)('0' + (value / 10) % 10);
-  string[startPos + 0] = (byte)('0' + (value / 100) % 10);
-}
-
-void incrementIterCounter() {
-  byte counterValue;
-  int len = storageGetValueLength(recursiveIterationCounterKey);
-  if (len == 0) {
-    counterValue = 1;
-    storageStore(recursiveIterationCounterKey, &counterValue, 1);
-  } else {
-    storageLoad(recursiveIterationCounterKey, &counterValue); 
-    counterValue = counterValue + 1;
-    storageStore(recursiveIterationCounterKey, &counterValue, 1);
-  }
-}
-
-void incrementBigIntCounter() {
-  bigIntSetInt64(42, 1);
-  bigIntAdd(bigIntIterationCounterID, bigIntIterationCounterID, 42);
-}
-
-void finishResult(int result) {
-	if (result == 0) {
-		byte message[] = "succ";
-		finish(message, 4);
-	}
-	if (result == 1) {
-		byte message[] = "fail";
-		finish(message, 4);
-	}
-	if (result != 0 && result != 1) {
-		byte message[] = "unkn";
-		finish(message, 4);
-	}
 }
